+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.3.3] - 2019-12-12
+
+Fix metadata reporting for editor tests
+Fix configuration fields in json report
 
 ## [1.3.2] - 2019-11-27